30min time add fix

+++ backend/bookings/views.py
@@ -148,7 +148,7 @@ class BookingAvailabilityViewSet(viewsets.ViewSet):
         try:
             start_date = datetime.strptime(start_date, '%Y-%m-%d').date()
             end_date = datetime.strptime(end_date, '%Y-%m-%d').date()
-            duration_hours = int(duration_hours)
+            duration_hours = float(duration_hours)
         except (ValueError, TypeError):
             return Response(
                 {'error': 'Invalid date format'},
